feat: add navigation btn
